Update pco_pixelfly_usb.py
Deep and confusing bug resolved! "dll.get_camera_name()" must be passed a pointer to a "C char" datatype (c_char_p) that is 40 bytes long. The Python level bug was that len(b''*40) == 0 (not 40) so a space character is needed to give the correct length (i.e. len(b' '*40) == 40). Despite this Python level error the dll would return the correct name (with no error) and seemingly function normally. However the resulting (seemingly unrelated) failure was a Python (or OS level!) crash from a read/write error after approximately ~17000 "dll.allocate_buffer()" and "dll.free_buffer()" cycles! (just over ~1000 cycles if all 16 buffers were allocated/freed). The origin of this failure is not understood and is buried somewhere in the PCO dll.